add blueprint
