+++ README.md
@@ -1,0 +1,17 @@
+##pageOs_v0.1 
+
+PageOs_v0.1
+
+    ├── bokchoy_pages  # page object files
+    ├── faker_data     # MOCK data
+    ├── resources      # Flask-blueprint
+    ├── similar_analysis # web page similar analysis
+    ├── test-output # .dot/.png files
+    ├── test_gen # testcases generate
+    ├── uploads # save upload file
+    
+   画图命令
+```dot -Tpng petclinic.txt -o petclinic.png ```
+
+RTED_v1.2.jar # Calculation tree distance
+page_generator.sh # call selenium page object generator
