--- create_db.py
@@ -1,11 +1,0 @@
-# -*- encoding: utf-8 -*-
-"""
-@File    : create_db.py
-@Time    : 2020/2/13 10:54 上午
-@Author  : zhengjiani
-@Email   : 936089353@qq.com
-@Software: PyCharm
-"""
-from app import db
-db.drop_all()
-db.create_all()
