--- petclinic/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-# @Time    : 2019/6/12 8:36
-# @Author  : zhengjiani
-# @Software: PyCharm
-# @Blog    ：https://zhengjiani.github.io/
