--- petclinic/selenium_approach/selenium_ori/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-# @Time    : 2019/6/4 17:31
-# @Author  : zhengjiani
-# @Software: PyCharm
-# @Blog    ：https://zhengjiani.github.io/
