--- petclinic/selenium_approach/selenium_ori/test.py
@@ -1,26 +1,0 @@
-# -*- coding: utf-8 -*-
-# @Time    : 2019/6/4 17:24
-# @Author  : zhengjiani
-# @Software: PyCharm
-# @Blog    ：https://zhengjiani.github.io/
-import unittest
-from selenium import webdriver
-from . import page
-class PythonOrgSearch(unittest.TestCase):
-    def setUp(self):
-        self.driver = webdriver.Chrome()
-        self.driver.get("http://www.python.org")
-
-    def test_search_in_python_org(self):
-        main_page = page.MainPage(self.driver)
-        assert main_page.is_title_matches()
-        main_page.search_text_element = "pycon"
-        main_page.click_go_button()
-        search_results_page = page.SearchResultsPage(self.driver)
-        assert search_results_page.is_results_found()
-
-    def tearDown(self):
-        self.driver.close()
-
-if __name__ == "__main__":
-    unittest.main()
