--- petclinic/selenium_approach/templates/testcase.j2
@@ -1,14 +1,0 @@
-{% extends 'base.j2' %}
-{%  block testcase -%}
-    {{ super() }}
-    {% for current_page,tc_name in dic_graph.items() %}
-        {% for func_name,params in dic_params.items() %}
-    def test_{{ tc_name }}(self):
-        driver = self.driver
-        driver.get('{{ url }}')
-        current_page = {{ current_page }}(driver)
-        current_page.{{ func_name }}{{ params }}
-        time.sleep(2)
-        {% endfor %}
-    {% endfor %}
-{%- endblock %}
