+++ test_gen/test_pet.py
@@ -1,12 +1,8 @@
 # -*- coding: utf-8 -*-
-# @Time    : 2019/6/9 8:57
-# @Author  : zhengjiani
-# @Software: PyCharm
-# @Blog    ：https://zhengjiani.github.io/
 import os
 import unittest
 from bok_choy.web_app_test import WebAppTest
-from .pet_page import *
+from bokchoy_pages.pet_page import *
 class TestPet(WebAppTest):
     def setUp(self):
         super(TestPet, self).setUp()
