--- utils/README.md
@@ -1,2 +1,0 @@
-画图命令
-```dot -Tpng petclinic.txt -o petclinic.png ```
