--- utils/petclinic.txt
@@ -1,56 +1,0 @@
-digraph G {
-  HomeContainerPage;
-  AddEditVisitContainerPage;
-  AddEditPetContainerPage;
-  VeterinariansContainerPage;
-  OwnerInformationContainerPage;
-  AddEditOwnerContainerPage;
-  OwnersListContainerPage;
-  HomeContainerPage -> HomeContainerPage [ label="goToHomeThroughLogoHomeContainerPage()" ];
-  HomeContainerPage -> HomeContainerPage [ label="goToHomeThroughNavbarHomeContainerPage()" ];
-  HomeContainerPage -> OwnersListContainerPage [ label="goToOwnersListHomeContainerPage()" ];
-  HomeContainerPage -> VeterinariansContainerPage [ label="goToVeterinariansHomeContainerPage()" ];
-  HomeContainerPage -> AddEditOwnerContainerPage [ label="registerNewOwnerHomeContainerPage()" ];
-  AddEditVisitContainerPage -> OwnerInformationContainerPage [ label="addVisitAddEditVisitContainerPage(date,visitDescription)" ];
-  AddEditVisitContainerPage -> HomeContainerPage [ label="goToHomeThroughLogoAddEditVisitContainerPage()" ];
-  AddEditVisitContainerPage -> HomeContainerPage [ label="goToHomeThroughNavbarAddEditVisitContainerPage()" ];
-  AddEditVisitContainerPage -> OwnersListContainerPage [ label="goToOwnersListAddEditVisitContainerPage()" ];
-  AddEditVisitContainerPage -> VeterinariansContainerPage [ label="goToVeterinariansAddEditVisitContainerPage()" ];
-  AddEditVisitContainerPage -> AddEditOwnerContainerPage [ label="registerNewOwnerAddEditVisitContainerPage()" ];
-  AddEditPetContainerPage -> OwnerInformationContainerPage [ label="addNewPetAddEditPetContainerPage(petName,date,petType)" ];
-  AddEditPetContainerPage -> AddEditPetContainerPage [ label="addNewPetWithLongNameAddEditPetContainerPage(petName,date,petType)" ];
-  AddEditPetContainerPage -> HomeContainerPage [ label="goToHomeThroughLogoAddEditPetContainerPage()" ];
-  AddEditPetContainerPage -> HomeContainerPage [ label="goToHomeThroughNavbarAddEditPetContainerPage()" ];
-  AddEditPetContainerPage -> OwnersListContainerPage [ label="goToOwnersListAddEditPetContainerPage()" ];
-  AddEditPetContainerPage -> VeterinariansContainerPage [ label="goToVeterinariansAddEditPetContainerPage()" ];
-  AddEditPetContainerPage -> AddEditOwnerContainerPage [ label="registerNewOwnerAddEditPetContainerPage()" ];
-  VeterinariansContainerPage -> HomeContainerPage [ label="goToHomeThroughLogoVeterinariansContainerPage()" ];
-  VeterinariansContainerPage -> HomeContainerPage [ label="goToHomeThroughNavbarVeterinariansContainerPage()" ];
-  VeterinariansContainerPage -> OwnersListContainerPage [ label="goToOwnersListVeterinariansContainerPage()" ];
-  VeterinariansContainerPage -> VeterinariansContainerPage [ label="goToVeterinariansVeterinariansContainerPage()" ];
-  VeterinariansContainerPage -> AddEditOwnerContainerPage [ label="registerNewOwnerVeterinariansContainerPage()" ];
-  OwnerInformationContainerPage -> AddEditPetContainerPage [ label="goToAddNewPetOwnerInformationContainerPage()" ];
-  OwnerInformationContainerPage -> AddEditOwnerContainerPage [ label="goToEditOwnerOwnerInformationContainerPage()" ];
-  OwnerInformationContainerPage -> AddEditPetContainerPage [ label="goToEditPetThroghEditLinkOwnerInformationContainerPage(petName)" ];
-  OwnerInformationContainerPage -> AddEditPetContainerPage [ label="goToEditPetThroughNameOwnerInformationContainerPage(petName)" ];
-  OwnerInformationContainerPage -> HomeContainerPage [ label="goToHomeThroughLogoOwnerInformationContainerPage()" ];
-  OwnerInformationContainerPage -> HomeContainerPage [ label="goToHomeThroughNavbarOwnerInformationContainerPage()" ];
-  OwnerInformationContainerPage -> OwnersListContainerPage [ label="goToOwnersListOwnerInformationContainerPage()" ];
-  OwnerInformationContainerPage -> VeterinariansContainerPage [ label="goToVeterinariansOwnerInformationContainerPage()" ];
-  OwnerInformationContainerPage -> AddEditVisitContainerPage [ label="goToVisitsOwnerInformationContainerPage(petName)" ];
-  OwnerInformationContainerPage -> AddEditOwnerContainerPage [ label="registerNewOwnerOwnerInformationContainerPage()" ];
-  AddEditOwnerContainerPage -> HomeContainerPage [ label="goToHomeThroughLogoAddEditOwnerContainerPage()" ];
-  AddEditOwnerContainerPage -> HomeContainerPage [ label="goToHomeThroughNavbarAddEditOwnerContainerPage()" ];
-  AddEditOwnerContainerPage -> OwnersListContainerPage [ label="goToOwnersListAddEditOwnerContainerPage()" ];
-  AddEditOwnerContainerPage -> VeterinariansContainerPage [ label="goToVeterinariansAddEditOwnerContainerPage()" ];
-  AddEditOwnerContainerPage -> AddEditOwnerContainerPage [ label="registerNewOwnerAddEditOwnerContainerPage()" ];
-  AddEditOwnerContainerPage -> OwnersListContainerPage [ label="registerOwnerAddEditOwnerContainerPage(firstName,lastName,address,city,telephone)" ];
-  AddEditOwnerContainerPage -> AddEditOwnerContainerPage [ label="registerOwnerWithLongFirstNameAddEditOwnerContainerPage(firstName,lastName,address,city,telephone)" ];
-  OwnersListContainerPage -> HomeContainerPage [ label="goToHomeThroughLogoOwnersListContainerPage()" ];
-  OwnersListContainerPage -> HomeContainerPage [ label="goToHomeThroughNavbarOwnersListContainerPage()" ];
-  OwnersListContainerPage -> OwnerInformationContainerPage [ label="goToOwnerInformationOwnersListContainerPage(firstName,lastName)" ];
-  OwnersListContainerPage -> OwnersListContainerPage [ label="goToOwnersListOwnersListContainerPage()" ];
-  OwnersListContainerPage -> VeterinariansContainerPage [ label="goToVeterinariansOwnersListContainerPage()" ];
-  OwnersListContainerPage -> AddEditOwnerContainerPage [ label="registerNewOwnerOwnersListContainerPage()" ];
-  OwnersListContainerPage -> OwnersListContainerPage [ label="searchFilterOwnersListContainerPage(firstName,lastName)" ];
-}
